polling w/ grid updates

+++ server/src/lib/database/migrations/1_create_pixels.ts
@@ -39,7 +39,7 @@ export const createPixels = async (knex: Knex) => {
 
             table.integer("row").unsigned();
             table.integer("column").unsigned();
-            table.integer("color").unsigned();
+            table.bigInteger("color").unsigned();
             table.timestamp("created_at");
 
             table.index(["grid_id", "created_at"]);
